oracle/requirements.txt
web3==6.15.1
python-dotenv==1.0.1
requests==2.31.0
pycoingecko==3.1.0
openai==1.12.0
